Move the colonies files to their own class.
This'll be useful when resolving ColonyDealer serialization.
